pyproject.toml
[tool.poetry]
name = "things"
version = "0.1.0"
description = ""
authors = ["LJlkdskdjflsa <liliangjya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
Django = "^4.1.2"
psycopg2-binary = "^2.9.5"
django-treenode = "^0.18.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
flake8 = "^5.0.4"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
